Create Part6 - Menu
The gourmet restaurant 'Unicafe' on the Kumpula campus of the University of Helsinki needs a new menu. The chef knows about programming and wants a computer system to manage the menu. In this assignment, we'll implement the heart of the system, the Menu class.

The exercise template comes with a Main class that you can use to test the menu. For the implementation of the menu, you'll have the following boilerplate code:

The menu object has an ArrayList as an instance variable to store the names of the dishes on the menu. The menu should provide the following methods:

public void addMeal(String meal) adds a meal to the menu. If the meal is already on the list, it should not be added again.

public void printMeals() prints the meals.

public void clearMenu() clears the menu.